requirements.txt
asgiref==3.6.0
click==8.1.3
colorama==0.4.6
Django==4.2.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.0
psycopg2-binary==2.9.6
PyJWT==2.7.0
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
tomli==2.0.1
tzdata==2023.3
